Fix deadlock in mq_send
If O_NONBLOCK is set, mq_send now relinquishes ownership of mqd, and
waits until a message slot becomes free. mq_send sleeps for
mq_attr.mq_sleepdur microseconds. On wake, the queue is checked for
free message slots and a message is sent, if possible.

struct mqd now holds the private attributes for the mqueue. Only the
shared information is contained in struct mqueue.