Change how queues are locked.
mq_open: First create and acquire a mutex, if the queue is not private.
Create the private lock last, and do not lock it.

mq_close: Do not lock the mutex, but only the private lock, as no
changes are made to the queue; only the descriptor is put into the free
list.

The rest of the functions require both locks, if present to be acquired
by the calling thread.